+++ iivs_cardio/common/pipeline/__init__.py
@@ -10,6 +10,7 @@ __all__ = (
     "FrameBranch",
     "FrameWriter",
     "Hook",
+    "ItemVerdict",
     "Named",
     "PresentPolicy",
     "Releasable",
@@ -69,4 +70,4 @@ from iivs_cardio.common.pipeline.frames import (
     FrameBranch,
     FrameWriter,
 )
-from iivs_cardio.common.pipeline.run import Releasable, StageRun
+from iivs_cardio.common.pipeline.run import ItemVerdict, Releasable, StageRun